Time: 0 ms (100%), Space: 23 MB (10.55%) - LeetHub

+++ 0026-remove-duplicates-from-sorted-array/0026-remove-duplicates-from-sorted-array.cpp
@@ -25,11 +25,11 @@
 class Solution {
 public:
     int removeDuplicates(vector<int>& nums) {
-        int sizeNums = nums.size();
+        short sizeNums = nums.size();
         int *indcies=new int[sizeNums];
         fill_n(indcies, sizeNums, -1);
         short start = 0;
-        for (int i = sizeNums - 1; i > 0; i--)
+        for (short i = sizeNums - 1; i > 0; i--)
         {
             if (nums[i] == nums[i - 1])
             {
@@ -37,7 +37,7 @@ public:
             }
 
         }
-        for (int z = 0; z <sizeNums ; z++)
+        for (short z = 0; z <sizeNums ; z++)
         {
             if (indcies[z] >= 0)
                 nums.erase(nums.begin() + indcies[z]);
